backend/pyproject.toml
[project]
name = "regimeforge"
version = "0.1.0"
description = "Regime-aware reinforcement learning workbench for synthetic trading research"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "RegimeForge contributors" },
]
keywords = [
  "reinforcement-learning",
  "algorithmic-trading",
  "mixture-of-experts",
  "market-regimes",
  "research-tooling",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Office/Business :: Financial",
]
dependencies = [
  "filelock>=3.16",
  "fastapi>=0.116",
  "jinja2>=3.1",
  "matplotlib>=3.9",
  "numpy>=1.26",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "rich>=13.9",
  "scikit-learn>=1.5",
  "scipy>=1.13",
  "tensorboard>=2.18",
  "torch>=2.7",
  "uvicorn>=0.35",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.15",
]

[project.scripts]
regimeforge-tui = "regime_lens.tui:main"
regimeforge-experiments = "regime_lens.run_experiments:main"
regimeforge-dashboard = "regime_lens.web:main"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["regime_lens*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E9", "F63", "F7", "F82"]
